+++ epic_app/tests/serializers/test_summary_serializer.py
@@ -8,11 +8,12 @@ from rest_framework.test import APIRequestFactory
 from epic_app.models.epic_answers import EvolutionAnswer
 from epic_app.models.epic_questions import EvolutionChoiceType, EvolutionQuestion
 from epic_app.models.epic_user import EpicOrganization, EpicUser
+from epic_app.models.models import Program
 from epic_app.serializers.summary_serializer import (
     SummaryEvolutionSerializer,
     SummaryOrganizationEvolutionSerializer,
 )
-from epic_app.tests import django_postgresql_db
+from epic_app.tests import django_postgresql_db, test_data_dir
 from epic_app.tests.epic_db_fixture import epic_test_db
 
 
@@ -24,6 +25,7 @@ def get_serializer():
     return {
         "request": Request(request),
         "users": EpicUser.objects.all(),
+        "organizations": EpicOrganization.objects.all(),
     }
 
 
@@ -55,24 +57,36 @@ def rest_framework_url_fixture(epic_test_db: pytest.fixture):
 class TestSummaryEvolutionSerializer:
     def test_summary_evolution_to_representation(self):
         # 1. Define test data and expectations.
-        evo_q = EvolutionQuestion.objects.all().first()
-        evo_q_ans = (
-            EvolutionAnswer.objects.filter(question=evo_q)
-            .all()
-            .values_list("selected_choice", flat=True)
-        )
-        evo_q_ans_avg = round(mean(map(EvolutionChoiceType.to_int, evo_q_ans)), 2)
+        test_p: Program = EvolutionQuestion.objects.all().first().program
+        org_avg = []
+        for e_org in EpicOrganization.objects.all():
+            user_avg = []
+            for e_user in e_org.organization_users.all():
+                _answers = (
+                    EvolutionAnswer.objects.filter(
+                        user=e_user, question__in=test_p.questions.all()
+                    )
+                    .all()
+                    .values_list("selected_choice", flat=True)
+                )
+                if not _answers:
+                    _answers = [0]
+                user_avg.append(mean(list(map(EvolutionChoiceType.to_int, _answers))))
+            org_avg.append(mean(user_avg))
+        program_avg = round(mean(org_avg), 2)
 
         # 2. Run test
         represented_data = SummaryEvolutionSerializer(
             context=serializer_context
-        ).to_representation(evo_q)
+        ).to_representation(test_p)
 
         # 3. Verify final expectations.
         assert represented_data == {
-            "id": evo_q.pk,
-            "title": evo_q.title,
-            "average": evo_q_ans_avg,
+            "id": test_p.pk,
+            "area": test_p.group.area.name,
+            "group": test_p.group.name,
+            "program": test_p.name,
+            "average": program_avg,
         }
 
 
@@ -83,7 +97,7 @@ class TestSummaryOrganizationEvolutionSerializer:
             context=serializer_context
         ).to_representation(EpicOrganization.objects.all().first())
         assert set(represented_data.keys()) == set(
-            ["id", "organization", "evolution_questions"]
+            ["id", "organization", "evolution_summary"]
         )
-        assert isinstance(represented_data["evolution_questions"], list)
-        assert isinstance(represented_data["evolution_questions"][0], dict)
+        assert isinstance(represented_data["evolution_summary"], list)
+        assert isinstance(represented_data["evolution_summary"][0], dict)